refactor(categories): create new hook and update hooks
